knative-image-processing-workflow/functions/processor/requirements.txt
parliament-functions==0.1.0
functions-framework==3.*
cloudevents==1.9.0
minio==7.1.15
pillow==10.3.0
requests==2.31.0
python-magic==0.4.27